proxy: Move setup into proxy.py
