+++ telepresence/proxy/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Datawire. All rights reserved.
+# Copyright 2020 Datawire. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,136 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import re
-from subprocess import CalledProcessError
 
-from telepresence.runner import Runner
-from telepresence.utilities import get_alternate_nameserver
+from .proxy import setup
+from .remote import RemoteInfo
 
-from .deployment import (
-    create_new_deployment, existing_deployment, existing_deployment_openshift,
-    supplant_deployment, swap_deployment_openshift
-)
-from .remote import RemoteInfo, get_remote_info
-
-
-def _dc_exists(runner: Runner, name: str) -> bool:
-    """
-    If we're using OpenShift Origin, we may be using a DeploymentConfig instead
-    of a Deployment. Return True if a dc exists with the given name.
-    """
-    # Need to use oc to manage DeploymentConfigs. The cluster needs to be
-    # running OpenShift as well. Check for both.
-    kube = runner.kubectl
-    if kube.command != "oc" or not kube.cluster_is_openshift:
-        return False
-    if ":" in name:
-        name, container = name.split(":", 1)
-    try:
-        runner.check_call(runner.kubectl("get", "dc/{}".format(name)))
-        return True
-    except CalledProcessError as exc:
-        runner.show(
-            "Failed to find OpenShift deploymentconfig {}:".format(name)
-        )
-        runner.show("  {}".format(str(exc.stderr)))
-        runner.show("Will try regular Kubernetes Deployment.")
-    return False
-
-
-def setup(runner: Runner, args):
-    """
-    Determine how the user wants to set up the proxy in the cluster.
-    """
-
-    # OpenShift doesn't support running as root:
-    if (
-        args.expose.has_privileged_ports()
-        and runner.kubectl.cluster_is_openshift
-    ):
-        raise runner.fail("OpenShift does not support ports <1024.")
-
-    # Check the service account, if present
-    if args.service_account:
-        try:
-            runner.check_call(
-                runner.kubectl("get", "serviceaccount", args.service_account)
-            )
-        except CalledProcessError as exc:
-            raise runner.fail(
-                "Check service account {} failed:\n{}".format(
-                    args.service_account, exc.stderr
-                )
-            )
-
-    # Figure out which operation the user wants
-    if args.deployment is not None:
-        # This implies --deployment
-        if _dc_exists(runner, args.deployment_arg):
-            operation = existing_deployment_openshift
-            deployment_type = "deploymentconfig"
-        else:
-            operation = existing_deployment
-            deployment_type = "deployment"
-
-    if args.new_deployment is not None:
-        # This implies --new-deployment
-        deployment_type = "deployment"
-        operation = create_new_deployment
-
-    if args.swap_deployment is not None:
-        # This implies --swap-deployment
-        if _dc_exists(runner, args.deployment_arg):
-            operation = swap_deployment_openshift
-            deployment_type = "deploymentconfig"
-        else:
-            operation = supplant_deployment
-            deployment_type = "deployment"
-
-    # minikube/minishift break DNS because DNS gets captured, sent to minikube,
-    # which sends it back to the DNS server set by host, resulting in a DNS
-    # loop... We've fixed that for most cases by setting a distinct name server
-    # for the proxy to use when making a new proxy pod, but that does not work
-    # for --deployment.
-    deployment_env = {}
-    if args.method == "vpn-tcp" and runner.kubectl.in_local_vm:
-        if args.operation == "deployment":
-            raise runner.fail(
-                "vpn-tcp method doesn't work with minikube/minishift when"
-                " using --deployment. Use --swap-deployment or"
-                " --new-deployment instead."
-            )
-        try:
-            deployment_env["TELEPRESENCE_NAMESERVER"] \
-                = get_alternate_nameserver()
-            if args.also_proxy:
-                proxy_names = []
-                for name in args.also_proxy:
-                    if not (
-                        re.search(r"[^\w.]", name)
-                        or re.match(r"^(?:\d+\.){3}\d+$", name)
-                    ):
-                        proxy_names.append(name)
-                if proxy_names:
-                    deployment_env["TELEPRESENCE_LOCAL_NAMES"] \
-                        = ",".join(proxy_names)
-
-        except Exception as exc:
-            raise runner.fail(
-                "Failed to find a fallback nameserver: {}".format(exc)
-            )
-
-    def start_proxy(runner_: Runner) -> RemoteInfo:
-        tel_deployment, run_id = operation(
-            runner_, args.deployment_arg, args.expose, deployment_env,
-            args.service_account
-        )
-        remote_info = get_remote_info(
-            runner,
-            tel_deployment,
-            deployment_type,
-            run_id=run_id,
-        )
-        return remote_info
-
-    return start_proxy
+__all__ = "RemoteInfo", "setup",
